+++ CHANGELOG.md
@@ -4,4 +4,6 @@ swagger-result
 ##Staged
  - Initial commit
 
+### Added
+ - preOutput function - additional setting and formatting for context